# docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  domain.DayInput:
    properties:
      created_at:
        type: number
    type: object
...
  domain.GetAllDaysResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.DayInput'
        type: array
      page-info:
        $ref: '#/definitions/domain.PaginationPage'
    type: object
...
  domain.GetAllWinnersCategoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Winners'
        type: array
      page-info:
        $ref: '#/definitions/domain.PaginationPage'
    type: object
...
  domain.PaginationPage:
    properties:
      count:
        type: integer
      page:
        type: integer
      pages:
        type: integer
    type: object
...
  domain.WinnerInput:
    properties:
      check_id:
        type: integer
      end_check_date:
        type: number
      end_register_date:
        type: number
      money_amount:
        type: integer
      partner_id:
        type: integer
      raffle_id:
        type: integer
      start_check_date:
        type: number
      start_register_date:
        type: number
    required:
    - check_id
    - raffle_id
    type: object
  domain.Winners:
    properties:
      check_id:
        type: integer
      id:
        type: integer
      is_winner:
        type: boolean
      phone_number:
        type: string
      raffle_date:
        type: number
      raffle_type:
        type: string
      user_name:
        type: string
    type: object
...
  v1.okResponse:
    properties:
      message:
        type: string
    type: object
...
  v1.response:
    properties:
      detail:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: API Server for  A Hundred to Five
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: A Hundred to Five
  version: "2.0"
paths:
...
  /days:
    get:
      consumes:
      - application/json
      description: gets all days
      operationId: getAllDays
      parameters:
      - in: query
        name: month
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetAllDaysResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - winners
...
  /members:
    get:
      consumes:
      - application/json
      description: gets all members
      operationId: getAllMembers
      parameters:
      - in: query
        name: raffle_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetAllWinnersCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - winners
  /months:
    get:
      consumes:
      - application/json
      description: gets all days
      operationId: getAllMonths
      parameters:
      - in: query
        minimum: 0
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetAllDaysResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - winners
...
  /winner:
    get:
      consumes:
      - application/json
      description: gets all winners
      operationId: get-all-winners
      parameters:
      - in: query
        name: date
        type: integer
      - in: query
        minimum: 0
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetAllWinnersCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - winners
    post:
      consumes:
      - application/json
      parameters:
      - description: winners
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/domain.WinnerInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.okResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - winners
securityDefinitions:
  User_Auth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"